Set prediction current day emotions deadlines fixed

+++ src/main/java/pl/skyterix/sadsky/user/domain/dto/UserDTO.java
@@ -22,7 +22,7 @@ public class UserDTO {
 
     private LocalDate birthDay;
 
-    private Short wakeHour;
+    private Integer wakeHour;
 
     @ToString.Exclude
     @EqualsAndHashCode.Exclude
